Configurado banco H2

+++ src/main/java/com/rafaelehlert/eventos/entities/Atividade.java
@@ -23,7 +23,7 @@ public class Atividade {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String nome;
 
@@ -48,18 +48,18 @@ public class Atividade {
     public Atividade() {
     }
 
-    public Atividade(Long id, String nome, String descricao, Double preco) {
+    public Atividade(Integer id, String nome, String descricao, Double preco) {
         this.id = id;
         this.nome = nome;
         this.descricao = descricao;
         this.preco = preco;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
