+++ src/main/java/com/rafaelehlert/eventos/entities/Participante.java
@@ -16,7 +16,7 @@ public class Participante {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String nome;
     private String email;
@@ -27,17 +27,17 @@ public class Participante {
     public Participante() {
     }
 
-    public Participante(Long id, String nome, String email) {
+    public Participante(Integer id, String nome, String email) {
         this.id = id;
         this.nome = nome;
         this.email = email;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
